Reverted default implementation of FsController.isExecutable(FsEntryName) to return false again.
First, while throwing new IOException(new UnsupportedOperationException()) as by the last changeset would be a possible alternative if this attribute is not supported by this controller, such behavior may surprise applications and might get misinterpreted as a true I/O failure. However, most applications are simply interested in the effective value for the attribute, regardless of any driver specific issues, so returning false as a default makes more sense.

Second, after all, an IOException should be an exception, not the standard case.

In TrueZIP 8, the return value of some methods may be changed from boolean to Boolean. This would allow to return null in order to signal that this attribute is not supported by this controller. An IOException as a result should still be reserved for exceptional situations, e.g. a RemoteException could get used to indicate a network connection problem.